+++ client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.12.7
 
+- Add support for the Light Client Tracking Service
 - Update `avail-light-core` to 1.2.2
 
 ## [1.12.6](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.6) - 2025-02-10
@@ -10,7 +11,6 @@
 
 ## [1.12.5](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.5) - 2025-01-23
 
-- Add support for the Light Client Tracking Service
 - Update `avail-light-core` to 1.2.0
 - Remove `ot_flush_block_interval` from configuration