+++ client/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.12.7
+## [1.12.7](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.7) - 2025-02-20
 
 - Add support for the Light Client Tracking Service
 - Update `avail-light-core` to 1.2.2